feat: The `AParserTree` class has been deprecated and replaced with the new `CST` class.
