+++ astars/__init__.py
@@ -1,11 +1,11 @@
-from .nodes     import ANode
+from .cst       import CSTNode
 from .parser    import AParser
 from .pruner    import APruner
 from .traverser import ATraverser
 from .traverser import AReverseTraverser
 from .tree      import AParseTree
 
-__all__ = [ "ANode", 
+__all__ = [ "CSTNode", 
             "AParser", 
             "APruner", 
             "ATraverser",
