--- astars/nodes/__init__.py
@@ -1,1 +1,0 @@
-from .node import ANode
