+++ astars/text/__init__.py
@@ -1,0 +1,1 @@
+from .text import SourceText
