+++ astars/text/text.py
@@ -1,0 +1,5 @@
+from dataclasses import dataclass
+
+@dataclass
+class SourceText:
+    source_code: str = ""
